% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{AnalysisResult}
\alias{NormativeReference}
\alias{PhysioBiomarker}
\alias{PhysioEvents}
\alias{PhysioExperiment}
\alias{addEvents}
\alias{appendProvenance}
\alias{applyMontage}
\alias{availableOperations}
\alias{availableReaders}
\alias{availableWriters}
\alias{biomarkerValue}
\alias{blandAltman}
\alias{cbindPhysio}
\alias{channelInfo}
\alias{channelInfo<-}
\alias{channelNames}
\alias{channelNames<-}
\alias{checkNA}
\alias{cohensD}
\alias{defaultAssay}
\alias{deprecate_physio}
\alias{dropChannels}
\alias{duration}
\alias{etaSquared}
\alias{eventQuery}
\alias{extractWindow}
\alias{fillEdgeNA}
\alias{filterType}
\alias{filterValue}
\alias{getChannelsByType}
\alias{getElectrodePositions}
\alias{getEvents}
\alias{getOperation}
\alias{getReader}
\alias{getReference}
\alias{getWriter}
\alias{handleNA}
\alias{hasNA}
\alias{icc}
\alias{logStep}
\alias{mdc}
\alias{nChannels}
\alias{nEvents}
\alias{naSummary}
\alias{percentPredicted}
\alias{physioPalette}
\alias{pickChannels}
\alias{provenance}
\alias{provenance<-}
\alias{rbindPhysio}
\alias{registerOperation}
\alias{registerReader}
\alias{registerWriter}
\alias{removeEvents}
\alias{renameChannels}
\alias{replaceNA}
\alias{resolveQuery}
\alias{resultType}
\alias{resultValue}
\alias{samplesToTime}
\alias{samplingRate}
\alias{samplingRate<-}
\alias{scale_color_physio}
\alias{scale_colour_physio}
\alias{scale_fill_physio}
\alias{sem}
\alias{setChannelTypes}
\alias{setChannelUnits}
\alias{setElectrodePositions}
\alias{setEvents}
\alias{setReference}
\alias{theme_physio}
\alias{timeIndex}
\alias{timeToSamples}
\alias{unregisterOperation}
\alias{unregisterReader}
\alias{unregisterWriter}
\alias{withProvenance}
\alias{zScore}
\alias{connectDatabase}
\alias{dbStats}
\alias{deleteExperiment}
\alias{disconnectDatabase}
\alias{initPhysioSchema}
\alias{isHDF5Backed}
\alias{listBIDSSessions}
\alias{listBIDSSubjects}
\alias{loadExperiment}
\alias{mapClinicalCodes}
\alias{queryExperiments}
\alias{readBDF}
\alias{readBIDS}
\alias{readBrainVision}
\alias{readCSV}
\alias{readClinicalMetadataCSV}
\alias{readEDF}
\alias{readElectrodePositionsCSV}
\alias{readEventsCSV}
\alias{readGDF}
\alias{readMAT}
\alias{readPhysio}
\alias{readPhysioHDF5}
\alias{realizeHDF5}
\alias{registerExperiment}
\alias{validateBIDS}
\alias{validateClinicalMetadata}
\alias{writeAssayHDF5}
\alias{writeBDF}
\alias{writeBIDS}
\alias{writeBrainVision}
\alias{writeCSV}
\alias{writeEDF}
\alias{writeElectrodePositionsCSV}
\alias{writeEventsCSV}
\alias{writeGDF}
\alias{writeMAT}
\alias{writePhysio}
\alias{writePhysioHDF5}
\alias{applyPipeline}
\alias{assaySamplingRates}
\alias{baselineCorrect}
\alias{butterworthFilter}
\alias{classifyICAComponents}
\alias{cleanData}
\alias{createPipeline}
\alias{decimate}
\alias{detectArtifacts}
\alias{detectBadChannels}
\alias{detrendSignal}
\alias{detrendSignals}
\alias{filterSignals}
\alias{firFilter}
\alias{getCurrentReference}
\alias{icaDecompose}
\alias{icaRemove}
\alias{interpolate}
\alias{interpolateBadChannels}
\alias{isAverageReferenced}
\alias{notchFilter}
\alias{rejectBadEpochs}
\alias{removeBaseline}
\alias{removeICAComponents}
\alias{rereference}
\alias{resample}
\alias{runICA}
\alias{setAssaySamplingRate}
\alias{adjacencyMatrix}
\alias{anovaEpochs}
\alias{averageEpochs}
\alias{bandPower}
\alias{betweennessCentrality}
\alias{binarizeNetwork}
\alias{bootstrapCI}
\alias{clusterPermutationTest}
\alias{clusteringCoefficient}
\alias{coherence}
\alias{connectivityMatrix}
\alias{correctPValues}
\alias{correlationMatrix}
\alias{crossSpectrum}
\alias{effectSize}
\alias{eigenvectorCentrality}
\alias{epochData}
\alias{epochSliding}
\alias{epochTimes}
\alias{fftSignals}
\alias{findSignificantWindows}
\alias{globalEfficiency}
\alias{grandAverage}
\alias{graphLaplacian}
\alias{hilbertTransform}
\alias{instantaneousAmplitude}
\alias{instantaneousPhase}
\alias{localEfficiency}
\alias{modularity}
\alias{nodeDegree}
\alias{pathLength}
\alias{pli}
\alias{plotAdjacencyMatrix}
\alias{plotDynamicConnectivity}
\alias{plotERP}
\alias{plotMultiChannel}
\alias{plotNetwork}
\alias{plotNetworkMetrics}
\alias{plotNetworkStability}
\alias{plotPSD}
\alias{plotSPM}
\alias{plotSignal}
\alias{plotSpectrogram}
\alias{plotTopomap}
\alias{plotTopomapSeries}
\alias{plv}
\alias{slidingWindowConnectivity}
\alias{smallWorldness}
\alias{spectralClustering}
\alias{spectralDecomposition}
\alias{spectrogram}
\alias{spmAnova}
\alias{spmPairedTTest}
\alias{spmTTest}
\alias{tTestEpochs}
\alias{temporalStability}
\alias{thresholdNetwork}
\alias{wPLI}
\alias{waveletTransform}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{PhysioAnalysis}{\code{\link[PhysioAnalysis]{adjacencyMatrix}}, \code{\link[PhysioAnalysis]{anovaEpochs}}, \code{\link[PhysioAnalysis]{averageEpochs}}, \code{\link[PhysioAnalysis]{bandPower}}, \code{\link[PhysioAnalysis]{betweennessCentrality}}, \code{\link[PhysioAnalysis]{binarizeNetwork}}, \code{\link[PhysioAnalysis]{bootstrapCI}}, \code{\link[PhysioAnalysis]{clusteringCoefficient}}, \code{\link[PhysioAnalysis]{clusterPermutationTest}}, \code{\link[PhysioAnalysis]{coherence}}, \code{\link[PhysioAnalysis]{connectivityMatrix}}, \code{\link[PhysioAnalysis]{correctPValues}}, \code{\link[PhysioAnalysis]{correlationMatrix}}, \code{\link[PhysioAnalysis]{crossSpectrum}}, \code{\link[PhysioAnalysis]{effectSize}}, \code{\link[PhysioAnalysis]{eigenvectorCentrality}}, \code{\link[PhysioAnalysis]{epochData}}, \code{\link[PhysioAnalysis]{epochSliding}}, \code{\link[PhysioAnalysis]{epochTimes}}, \code{\link[PhysioAnalysis]{fftSignals}}, \code{\link[PhysioAnalysis]{findSignificantWindows}}, \code{\link[PhysioAnalysis]{globalEfficiency}}, \code{\link[PhysioAnalysis]{grandAverage}}, \code{\link[PhysioAnalysis]{graphLaplacian}}, \code{\link[PhysioAnalysis]{hilbertTransform}}, \code{\link[PhysioAnalysis]{instantaneousAmplitude}}, \code{\link[PhysioAnalysis]{instantaneousPhase}}, \code{\link[PhysioAnalysis]{localEfficiency}}, \code{\link[PhysioAnalysis]{modularity}}, \code{\link[PhysioAnalysis]{nodeDegree}}, \code{\link[PhysioAnalysis]{pathLength}}, \code{\link[PhysioAnalysis]{pli}}, \code{\link[PhysioAnalysis]{plotAdjacencyMatrix}}, \code{\link[PhysioAnalysis]{plotDynamicConnectivity}}, \code{\link[PhysioAnalysis]{plotERP}}, \code{\link[PhysioAnalysis]{plotMultiChannel}}, \code{\link[PhysioAnalysis]{plotNetwork}}, \code{\link[PhysioAnalysis]{plotNetworkMetrics}}, \code{\link[PhysioAnalysis]{plotNetworkStability}}, \code{\link[PhysioAnalysis]{plotPSD}}, \code{\link[PhysioAnalysis]{plotSignal}}, \code{\link[PhysioAnalysis]{plotSpectrogram}}, \code{\link[PhysioAnalysis]{plotSPM}}, \code{\link[PhysioAnalysis]{plotTopomap}}, \code{\link[PhysioAnalysis]{plotTopomapSeries}}, \code{\link[PhysioAnalysis]{plv}}, \code{\link[PhysioAnalysis]{slidingWindowConnectivity}}, \code{\link[PhysioAnalysis]{smallWorldness}}, \code{\link[PhysioAnalysis]{spectralClustering}}, \code{\link[PhysioAnalysis]{spectralDecomposition}}, \code{\link[PhysioAnalysis]{spectrogram}}, \code{\link[PhysioAnalysis]{spmAnova}}, \code{\link[PhysioAnalysis]{spmPairedTTest}}, \code{\link[PhysioAnalysis]{spmTTest}}, \code{\link[PhysioAnalysis]{temporalStability}}, \code{\link[PhysioAnalysis]{thresholdNetwork}}, \code{\link[PhysioAnalysis]{tTestEpochs}}, \code{\link[PhysioAnalysis]{waveletTransform}}, \code{\link[PhysioAnalysis]{wPLI}}}

  \item{PhysioExperiment}{\code{\link[PhysioExperiment]{addEvents}}, \code{\link[PhysioExperiment]{AnalysisResult}}, \code{\link[PhysioExperiment]{appendProvenance}}, \code{\link[PhysioExperiment]{applyMontage}}, \code{\link[PhysioExperiment:registerOperation]{availableOperations}}, \code{\link[PhysioExperiment:registerReader]{availableReaders}}, \code{\link[PhysioExperiment:registerWriter]{availableWriters}}, \code{\link[PhysioExperiment:AnalysisResult-accessors]{biomarkerValue}}, \code{\link[PhysioExperiment]{blandAltman}}, \code{\link[PhysioExperiment]{cbindPhysio}}, \code{\link[PhysioExperiment]{channelInfo}}, \code{\link[PhysioExperiment:channelInfo-set]{channelInfo<-}}, \code{\link[PhysioExperiment]{channelNames}}, \code{\link[PhysioExperiment:channelNames-set]{channelNames<-}}, \code{\link[PhysioExperiment]{checkNA}}, \code{\link[PhysioExperiment]{cohensD}}, \code{\link[PhysioExperiment]{defaultAssay}}, \code{\link[PhysioExperiment]{deprecate_physio}}, \code{\link[PhysioExperiment]{dropChannels}}, \code{\link[PhysioExperiment]{duration}}, \code{\link[PhysioExperiment]{etaSquared}}, \code{\link[PhysioExperiment]{eventQuery}}, \code{\link[PhysioExperiment]{extractWindow}}, \code{\link[PhysioExperiment]{fillEdgeNA}}, \code{\link[PhysioExperiment]{filterType}}, \code{\link[PhysioExperiment]{filterValue}}, \code{\link[PhysioExperiment]{getChannelsByType}}, \code{\link[PhysioExperiment]{getElectrodePositions}}, \code{\link[PhysioExperiment]{getEvents}}, \code{\link[PhysioExperiment:registerOperation]{getOperation}}, \code{\link[PhysioExperiment:registerReader]{getReader}}, \code{\link[PhysioExperiment]{getReference}}, \code{\link[PhysioExperiment:registerWriter]{getWriter}}, \code{\link[PhysioExperiment]{handleNA}}, \code{\link[PhysioExperiment]{hasNA}}, \code{\link[PhysioExperiment]{icc}}, \code{\link[PhysioExperiment]{logStep}}, \code{\link[PhysioExperiment]{mdc}}, \code{\link[PhysioExperiment]{naSummary}}, \code{\link[PhysioExperiment]{nChannels}}, \code{\link[PhysioExperiment]{nEvents}}, \code{\link[PhysioExperiment]{NormativeReference}}, \code{\link[PhysioExperiment:NormativeReference-compare]{percentPredicted}}, \code{\link[PhysioExperiment]{PhysioBiomarker}}, \code{\link[PhysioExperiment]{PhysioEvents}}, \code{\link[PhysioExperiment]{PhysioExperiment}}, \code{\link[PhysioExperiment]{physioPalette}}, \code{\link[PhysioExperiment]{pickChannels}}, \code{\link[PhysioExperiment]{provenance}}, \code{\link[PhysioExperiment:provenance]{provenance<-}}, \code{\link[PhysioExperiment]{rbindPhysio}}, \code{\link[PhysioExperiment]{registerOperation}}, \code{\link[PhysioExperiment]{registerReader}}, \code{\link[PhysioExperiment]{registerWriter}}, \code{\link[PhysioExperiment]{removeEvents}}, \code{\link[PhysioExperiment]{renameChannels}}, \code{\link[PhysioExperiment]{replaceNA}}, \code{\link[PhysioExperiment]{resolveQuery}}, \code{\link[PhysioExperiment:AnalysisResult-accessors]{resultType}}, \code{\link[PhysioExperiment:AnalysisResult-accessors]{resultValue}}, \code{\link[PhysioExperiment]{samplesToTime}}, \code{\link[PhysioExperiment]{samplingRate}}, \code{\link[PhysioExperiment:samplingRate]{samplingRate<-}}, \code{\link[PhysioExperiment]{scale_color_physio}}, \code{\link[PhysioExperiment:scale_color_physio]{scale_colour_physio}}, \code{\link[PhysioExperiment:scale_color_physio]{scale_fill_physio}}, \code{\link[PhysioExperiment]{sem}}, \code{\link[PhysioExperiment]{setChannelTypes}}, \code{\link[PhysioExperiment]{setChannelUnits}}, \code{\link[PhysioExperiment]{setElectrodePositions}}, \code{\link[PhysioExperiment]{setEvents}}, \code{\link[PhysioExperiment]{setReference}}, \code{\link[PhysioExperiment]{theme_physio}}, \code{\link[PhysioExperiment]{timeIndex}}, \code{\link[PhysioExperiment]{timeToSamples}}, \code{\link[PhysioExperiment:registerOperation]{unregisterOperation}}, \code{\link[PhysioExperiment:registerReader]{unregisterReader}}, \code{\link[PhysioExperiment:registerWriter]{unregisterWriter}}, \code{\link[PhysioExperiment]{withProvenance}}, \code{\link[PhysioExperiment:NormativeReference-compare]{zScore}}}

  \item{PhysioIO}{\code{\link[PhysioIO]{connectDatabase}}, \code{\link[PhysioIO]{dbStats}}, \code{\link[PhysioIO]{deleteExperiment}}, \code{\link[PhysioIO:connectDatabase]{disconnectDatabase}}, \code{\link[PhysioIO]{initPhysioSchema}}, \code{\link[PhysioIO]{isHDF5Backed}}, \code{\link[PhysioIO]{listBIDSSessions}}, \code{\link[PhysioIO]{listBIDSSubjects}}, \code{\link[PhysioIO]{loadExperiment}}, \code{\link[PhysioIO]{mapClinicalCodes}}, \code{\link[PhysioIO]{queryExperiments}}, \code{\link[PhysioIO]{readBDF}}, \code{\link[PhysioIO]{readBIDS}}, \code{\link[PhysioIO]{readBrainVision}}, \code{\link[PhysioIO]{readClinicalMetadataCSV}}, \code{\link[PhysioIO]{readCSV}}, \code{\link[PhysioIO]{readEDF}}, \code{\link[PhysioIO]{readElectrodePositionsCSV}}, \code{\link[PhysioIO]{readEventsCSV}}, \code{\link[PhysioIO]{readGDF}}, \code{\link[PhysioIO]{readMAT}}, \code{\link[PhysioIO:writePhysio]{readPhysio}}, \code{\link[PhysioIO]{readPhysioHDF5}}, \code{\link[PhysioIO]{realizeHDF5}}, \code{\link[PhysioIO]{registerExperiment}}, \code{\link[PhysioIO]{validateBIDS}}, \code{\link[PhysioIO]{validateClinicalMetadata}}, \code{\link[PhysioIO]{writeAssayHDF5}}, \code{\link[PhysioIO]{writeBDF}}, \code{\link[PhysioIO]{writeBIDS}}, \code{\link[PhysioIO]{writeBrainVision}}, \code{\link[PhysioIO]{writeCSV}}, \code{\link[PhysioIO]{writeEDF}}, \code{\link[PhysioIO]{writeElectrodePositionsCSV}}, \code{\link[PhysioIO]{writeEventsCSV}}, \code{\link[PhysioIO]{writeGDF}}, \code{\link[PhysioIO]{writeMAT}}, \code{\link[PhysioIO]{writePhysio}}, \code{\link[PhysioIO]{writePhysioHDF5}}}

  \item{PhysioPreprocess}{\code{\link[PhysioPreprocess]{applyPipeline}}, \code{\link[PhysioPreprocess]{assaySamplingRates}}, \code{\link[PhysioPreprocess]{baselineCorrect}}, \code{\link[PhysioPreprocess]{butterworthFilter}}, \code{\link[PhysioPreprocess]{classifyICAComponents}}, \code{\link[PhysioPreprocess]{cleanData}}, \code{\link[PhysioPreprocess]{createPipeline}}, \code{\link[PhysioPreprocess]{decimate}}, \code{\link[PhysioPreprocess]{detectArtifacts}}, \code{\link[PhysioPreprocess]{detectBadChannels}}, \code{\link[PhysioPreprocess]{detrendSignal}}, \code{\link[PhysioPreprocess]{detrendSignals}}, \code{\link[PhysioPreprocess]{filterSignals}}, \code{\link[PhysioPreprocess]{firFilter}}, \code{\link[PhysioPreprocess]{getCurrentReference}}, \code{\link[PhysioPreprocess]{icaDecompose}}, \code{\link[PhysioPreprocess]{icaRemove}}, \code{\link[PhysioPreprocess]{interpolate}}, \code{\link[PhysioPreprocess]{interpolateBadChannels}}, \code{\link[PhysioPreprocess]{isAverageReferenced}}, \code{\link[PhysioPreprocess]{notchFilter}}, \code{\link[PhysioPreprocess]{rejectBadEpochs}}, \code{\link[PhysioPreprocess]{removeBaseline}}, \code{\link[PhysioPreprocess]{removeICAComponents}}, \code{\link[PhysioPreprocess]{rereference}}, \code{\link[PhysioPreprocess]{resample}}, \code{\link[PhysioPreprocess]{runICA}}, \code{\link[PhysioPreprocess]{setAssaySamplingRate}}}
}}

